Move BatchPlayerImplementation into its own file.
