+++ src/esp/sim/BatchReplayRenderer.cpp
@@ -5,6 +5,7 @@
 #include "BatchReplayRenderer.h"
 
 #include <esp/gfx_batch/DepthUnprojection.h>
+#include <esp/sim/BatchPlayerImplementation.h>
 #include "esp/sensor/CameraSensor.h"
 
 #include <Corrade/Containers/GrowableArray.h>
@@ -15,7 +16,6 @@
 namespace esp {
 namespace sim {
 
-// clang-tidy you're NOT HELPING
 using namespace Mn::Math::Literals;  // NOLINT
 
 BatchReplayRenderer::BatchReplayRenderer(
@@ -46,170 +46,6 @@ BatchReplayRenderer::BatchReplayRenderer(
 
   theOnlySensorName_ = sensor.uuid;
   theOnlySensorProjection_ = sensor.projectionMatrix();
-
-  class BatchPlayerImplementation
-      : public gfx::replay::AbstractPlayerImplementation {
-   public:
-    BatchPlayerImplementation(gfx_batch::Renderer& renderer,
-                              Mn::UnsignedInt sceneId)
-        : renderer_{renderer}, sceneId_{sceneId} {}
-
-   private:
-    bool isSupportedRenderAsset(
-        const Corrade::Containers::StringView& filepath) {
-      // Primitives aren't directly supported in the Magnum batch renderer. See
-      // https://docs.google.com/document/d/1ngA73cXl3YRaPfFyICSUHONZN44C-XvieS7kwyQDbkI/edit#bookmark=id.yq39718gqbwz
-
-      const std::array<Corrade::Containers::StringView, 12> primNamePrefixes = {
-          "capsule3DSolid",     "capsule3DWireframe", "coneSolid",
-          "coneWireframe",      "cubeSolid",          "cubeWireframe",
-          "cylinderSolid",      "cylinderWireframe",  "icosphereSolid",
-          "icosphereWireframe", "uvSphereSolid",      "uvSphereWireframe"};
-
-      // primitive render asset filepaths start with one of the above prefixes.
-      // Examples: icosphereSolid_subdivs_1
-      // capsule3DSolid_hemiRings_4_cylRings_1_segments_12_halfLen_3.25_useTexCoords_false_useTangents_false
-      for (const auto& primNamePrefix : primNamePrefixes) {
-        if (filepath.size() < primNamePrefix.size()) {
-          continue;
-        }
-
-        if (filepath.prefix(primNamePrefix.size()) == primNamePrefix) {
-          return false;
-        }
-      }
-
-      return true;
-    }
-
-    gfx::replay::NodeHandle loadAndCreateRenderAssetInstance(
-        const esp::assets::AssetInfo& assetInfo,
-        const esp::assets::RenderAssetInstanceCreationInfo& creation) override {
-      // TODO anything to use creation.flags for?
-      // TODO is creation.lightSetupKey actually mapping to anything in the
-      //  replay file?
-
-      if (!isSupportedRenderAsset(creation.filepath)) {
-        ESP_WARNING() << "Unsupported render asset: " << creation.filepath;
-        return nullptr;
-      }
-
-      /* If no such name is known yet, add as a file */
-      if (!renderer_.hasNodeHierarchy(creation.filepath)) {
-        ESP_WARNING()
-            << creation.filepath
-            << "not found in any composite file, loading from the filesystem";
-
-        ESP_CHECK(
-            renderer_.addFile(creation.filepath,
-                              gfx_batch::RendererFileFlag::Whole |
-                                  gfx_batch::RendererFileFlag::GenerateMipmap),
-            "addFile failed for " << creation.filepath);
-        CORRADE_INTERNAL_ASSERT(renderer_.hasNodeHierarchy(creation.filepath));
-      }
-
-      return reinterpret_cast<gfx::replay::NodeHandle>(
-          renderer_.addNodeHierarchy(
-              sceneId_, creation.filepath,
-              /* Baking the initial scaling and coordinate frame into the
-                 transformation */
-              Mn::Matrix4::scaling(creation.scale ? *creation.scale
-                                                  : Mn::Vector3{1.0f}) *
-                  Mn::Matrix4::from(
-                      Mn::Quaternion{assetInfo.frame.rotationFrameToWorld()}
-                          .toMatrix(),
-                      {}))
-          /* Returning incremented by 1 because 0 (nullptr) is treated as an
-             error */
-          + 1);
-    }
-
-    void deleteAssetInstance(const gfx::replay::NodeHandle node) override {
-      // TODO actually remove from the scene instead of setting a zero scale
-      renderer_.transformations(
-          sceneId_)[reinterpret_cast<std::size_t>(node) - 1] =
-          Mn::Matrix4{Mn::Math::ZeroInit};
-    }
-
-    void deleteAssetInstances(
-        const std::unordered_map<gfx::replay::RenderAssetInstanceKey,
-                                 gfx::replay::NodeHandle>&) override {
-      renderer_.clear(sceneId_);
-    }
-
-    void setNodeTransform(const gfx::replay::NodeHandle node,
-                          const Mn::Vector3& translation,
-                          const Mn::Quaternion& rotation) override {
-      renderer_.transformations(
-          sceneId_)[reinterpret_cast<std::size_t>(node) - 1] =
-          Mn::Matrix4::from(rotation.toMatrix(), translation);
-    }
-
-    void setNodeTransform(const gfx::replay::NodeHandle node,
-                          const Mn::Matrix4& transform) override {
-      renderer_.transformations(
-          sceneId_)[reinterpret_cast<std::size_t>(node) - 1] = transform;
-    }
-
-    Mn::Matrix4 hackGetNodeTransform(
-        const gfx::replay::NodeHandle node) const override {
-      return renderer_.transformations(
-          sceneId_)[reinterpret_cast<std::size_t>(node) - 1];
-    }
-
-    void changeLightSetup(const esp::gfx::LightSetup& lights) override {
-      if (!renderer_.maxLightCount()) {
-        ESP_WARNING() << "Attempted to change" << lights.size()
-                      << "lights for scene" << sceneId_
-                      << "but the renderer is configured without lights";
-        return;
-      }
-
-      renderer_.clearLights(sceneId_);
-      for (std::size_t i = 0; i != lights.size(); ++i) {
-        const gfx::LightInfo& light = lights[i];
-        CORRADE_INTERNAL_ASSERT(light.model == gfx::LightPositionModel::Global);
-
-        const std::size_t nodeId = renderer_.addEmptyNode(sceneId_);
-
-        /* Clang Tidy, you're stupid, why do you say that "lightId" is not
-           initialized?! I'm initializing it right in the branches below, I
-           won't zero-init it just to "prevent bugs" because an accidentally
-           zero-initialized variable is *also* a bug, you know? Plus I have
-           range asserts in all functions so your "suggestions" are completely
-           unhelpful. */
-        std::size_t lightId;  // NOLINT
-        if (light.vector.w()) {
-          renderer_.transformations(sceneId_)[nodeId] =
-              Mn::Matrix4::translation(light.vector.xyz());
-          lightId = renderer_.addLight(sceneId_, nodeId,
-                                       gfx_batch::RendererLightType::Point);
-        } else {
-          /* The matrix will be partially NaNs if the light vector is in the
-             direction of the Y axis, but that's fine -- we only really care
-             about the Z axis direction, which is always the "target" vector
-             normalized. */
-          // TODO for more robustness use something that "invents" some
-          //  arbitrary orthogonal axes instead of the NaNs, once Magnum has
-          //  such utility
-          renderer_.transformations(sceneId_)[nodeId] =
-              Mn::Matrix4::lookAt({}, light.vector.xyz(), Mn::Vector3::yAxis());
-          lightId = renderer_.addLight(
-              sceneId_, nodeId, gfx_batch::RendererLightType::Directional);
-        }
-
-        renderer_.lightColors(sceneId_)[lightId] = light.color;
-        // TODO use gfx::getAmbientLightColor(lights) once it's not hardcoded
-        //  to an arbitrary value and once it's possible to change the ambient
-        //  factor in the renderer at runtime (and not just in
-        //  RendererConfiguration::setAmbientFactor())
-        // TODO range, once Habitat has that
-      }
-    }
-
-    gfx_batch::Renderer& renderer_;
-    Mn::UnsignedInt sceneId_;
-  };
 
   for (Mn::UnsignedInt i = 0; i != cfg.numEnvironments; ++i) {
     arrayAppend(
